requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==0.7
cssselect2==0.7.0
Django==4.2.7
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
fonttools==4.48.1
gobject==0.1.0
html5lib==1.1
idna==3.6
pdfkit==1.0.0
Pillow==10.1.0
pycparser==2.21
pydyf==0.8.0
PyPDF2==3.0.1
pyphen==0.14.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.0
weasyprint==56.1
webencodings==0.5.1
zopfli==0.2.3
